Extract query database execution partial
